Update README.md
The purpose of this assignment is to investigate the classification performance of neural networks. You will implement, train and evaluate your neural network models using Pytorch on the notMNIST dataset. The functions are provided and you will need to fill-in after theTODO comment instead of coding them from scratch. In particular, in the first part, you will implement a fully connected neural networks (FNN). In the second part, you will implement the model training loop. You will also be asked to answer several questions related to your implementations. You are encouraged to look up Pytorch documentation for useful utility functions, at: https://pytorch.org/docs/stable/index.html. You can also refer to the demo covered in the lecture. The notMNIST dataset is stored as notMNIST.npz inside in the assignment folder. We highly recommend students to look at the tutorial files for this assignment.

To avoid any potential installation issue, you are encouraged to develop your solution using Google Colab notebooks. It is highly recommended that you train the neural network using a GPU.

Unlike in the tutorial video which implemented convolutional neural networks, we simply im-plemented a fully-connected network for the image recognition task.

Requirements

In your implementations, please use the function prototype provided (i.e. name of the function, inputs and outputs) in the detailed instructions presented in the remainder of this document. We will be testing your code using a test function that will evoke the provided function prototype. If our testing file is unable to recognize the function prototype you have implemented, you can lose significant portion of your marks. In the assignment folder, the following files are included in thestarter code folder: